requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=1.5.0
matplotlib>=3.7.0
opencv-python>=4.7.0
pillow>=9.5.0
tqdm>=4.65.0
onnx>=1.14.0
onnxruntime>=1.15.0